Added search option to car.html
I added the search function to the car.html page and I removed the list of hard coded cars that was displayed on the page. I also made duplicated directories of fonts and images as I reqognized that the fonts and images weren't shoving on the page. With some debugging it seems that the program thinks the sources are stored in styles/[insert name of font or image]
So I created two new directories inside the style directory which is basically just a duplicate of the font and image folder as I didn't want to move them completley in case of messing up others worrk. Thus duplicating them seemed like the safer option.